methodology_page.py
"""
Methodology page for the ATC Training Progress Tracker.
Contains detailed methodology, technology stack, and benefits information.
"""
import streamlit as st
from PIL import Image


def display_methodology_page():
    """Display the Methodology page content."""
    
    # Page header
    st.markdown("# 🔬 Methodology")
    st.markdown("---")
    
    # Methodology Section
    st.markdown("## 🔬 Training Management Methodology")
    
    with st.expander("📊 Training Process Workflow", expanded=True):
        st.markdown("""
        ### Step-by-Step Process:
        
        1. **Trainee Onboarding**: 
           - Initial assessment and baseline competency evaluation
           - Training plan creation and goal setting
           - Schedule establishment
        
        2. **Continuous Monitoring**: 
           - Real-time competency tracking across all eight core areas
           - Session-by-session performance recording
           - Progress visualization through dashboards and charts
        
        3. **Training Plan Development**:
           - Collaborative discussions between trainee and training manager
           - AI-powered recommendations for personalized training paths
           - Activity scheduling and focus area identification
           - Regular plan reviews and adjustments
        
        4. **Performance Analysis**:
           - Comprehensive data collection from all training sessions
           - Trend analysis and pattern recognition
           - Comparative evaluation against cohort and standards
           - Strengths and weaknesses identification
        
        5. **AI-Powered Insights**:
           - Performance prediction and trajectory modeling
           - Personalized training recommendations
           - Automated summary generation
           - Intelligent competency gap analysis
        
        6. **Validation Assessment**:
           - Formal evaluation across all competency areas
           - Structured assessment forms and scoring
           - Certification readiness evaluation
           - Final validation and certification
        
        7. **Reporting & Documentation**: 
           - Automated report generation
           - Progress tracking documentation
           - Certification records
           - Historical performance archives
        """)
        
        # Add workflow flowchart illustration (placeholder)
        st.markdown("### 📊 Process Flow Illustration")
        st.info("Training workflow visualization: Dashboard → Training Plan → Performance Analysis → Validation → Certification")
    
    # Competency Framework
    st.markdown("## 🎯 Competency Framework")
    
    st.markdown("""
    The ATC Training Progress Tracker evaluates trainees across **eight core competency areas**:
    """)
    
    col1, col2 = st.columns(2)
    
    with col1:
        st.markdown("""
        ### 1. Communication
        Verbal and written communication skills, clarity, professionalism, and effectiveness 
        in conveying critical information under all conditions.
        
        ### 2. Situational Awareness
        Ability to maintain awareness of all aircraft, weather conditions, airspace status, 
        and potential conflicts or hazards.
        
        ### 3. Decision Making
        Quality and speed of decisions, risk assessment, prioritization, and ability to 
        make sound judgments under pressure.
        
        ### 4. Workload Management
        Capacity to handle multiple tasks, prioritize effectively, maintain performance 
        during peak traffic, and manage stress.
        """)
    
    with col2:
        st.markdown("""
        ### 5. Teamwork
        Collaboration with colleagues, information sharing, support provision, and 
        contribution to team effectiveness.
        
        ### 6. Technical Knowledge
        Understanding of regulations, procedures, equipment, airspace structure, and 
        theoretical concepts.
        
        ### 7. Procedures
        Adherence to standard operating procedures, consistency, and proper protocol 
        application in all scenarios.
        
        ### 8. Problem Solving
        Ability to identify issues, analyze problems, develop solutions, and handle 
        non-standard situations effectively.
        """)
    
    st.markdown("---")
    
    # Scoring System
    st.markdown("## 📊 Scoring & Evaluation System")
    
    col1, col2, col3 = st.columns(3)
    
    with col1:
        st.success("""
        ### ✅ Proficient (80-100%)
        - Consistently meets or exceeds standards
        - Demonstrates mastery
        - Requires minimal guidance
        - Ready for independent operation
        """)
    
    with col2:
        st.warning("""
        ### 🟡 Developing (70-79%)
        - Generally meets standards
        - Shows progress
        - Requires some supervision
        - On track for proficiency
        """)
    
    with col3:
        st.error("""
        ### 🔴 Needs Improvement (<70%)
        - Below expected standards
        - Requires focused training
        - Needs close supervision
        - Improvement plan required
        """)
    
    st.markdown("---")
    
    # Technology Stack Section
    st.markdown("## 💻 Technology Stack")
    
    tech_col1, tech_col2 = st.columns(2)
    
    with tech_col1:
        st.markdown("""
        ### Core Technologies:
        - **Python**: Primary programming language
        - **Streamlit**: Web application framework
        - **Pandas**: Data processing and analysis
        - **Plotly**: Interactive data visualization
        """)
    
    with tech_col2:
        st.markdown("""
        ### AI & Processing:
        - **OpenAI GPT-4o**: AI-powered insights and analysis
        - **Real-time Analytics**: Live performance tracking
        - **Automated Reporting**: Document generation
        - **Data Persistence**: Session state management
        """)
    
    # Benefits Section
    st.markdown("## ✨ Key Benefits")
    
    benefits_col1, benefits_col2, benefits_col3 = st.columns(3)
    
    with benefits_col1:
        st.markdown("""
        ### 🚀 Efficiency
        - Centralized training management
        - Automated tracking and reporting
        - Real-time progress visibility
        - Streamlined assessment processes
        """)
    
    with benefits_col2:
        st.markdown("""
        ### 🎯 Effectiveness
        - Personalized training paths
        - Data-driven decision making
        - Early intervention capabilities
        - Optimized learning outcomes
        """)
    
    with benefits_col3:
        st.markdown("""
        ### 📈 Insights
        - AI-powered recommendations
        - Performance predictions
        - Trend analysis
        - Comparative benchmarking
        """)
    
    st.markdown("---")
    
    # AI Integration
    st.markdown("## 🤖 AI Integration")
    
    st.markdown("""
    The system integrates artificial intelligence to enhance training management:
    
    **1. Performance Analysis**  
    AI analyzes trainee performance data to identify patterns, predict outcomes, and 
    highlight areas requiring attention.
    
    **2. Personalized Recommendations**  
    Based on individual learning patterns and competency gaps, the AI suggests tailored 
    training activities and schedules.
    
    **3. Automated Summaries**  
    AI generates comprehensive performance summaries, assessment reports, and progress 
    documentation automatically.
    
    **4. Predictive Modeling**  
    The system predicts certification readiness, estimates time to competency targets, 
    and identifies potential training challenges.
    """)
    
    st.info("""
    **Note**: AI features are designed to support and enhance human decision-making, 
    not replace the expertise of training managers and assessors.
    """)
    
    st.markdown("---")
    
    # Best Practices
    st.markdown("## 📚 Best Practices")
    
    st.markdown("""
    ### For Training Managers:
    - Conduct regular training plan reviews (weekly recommended)
    - Use AI recommendations as guidance, not directives
    - Document all assessment observations thoroughly
    - Provide timely, constructive feedback to trainees
    - Monitor trends across multiple trainees for program improvement
    
    ### For Trainees:
    - Review dashboard regularly to track progress
    - Be proactive in identifying areas needing focus
    - Engage actively in training plan discussions
    - Prepare for assessments by reviewing performance history
    - Use AI insights to guide self-directed learning
    
    ### For Administrators:
    - Ensure data accuracy and completeness
    - Regular system configuration reviews
    - Monitor AI model performance
    - Generate periodic program effectiveness reports
    - Maintain certification and validation standards
    """)

    
    impl_col1, impl_col2 = st.columns(2)
    
    with impl_col1:
        st.markdown("""
        ### Data Processing:
        - **CSV Validation**: Automatic format checking and error reporting
        - **Address Cleaning**: Standardisation of address formats
        - **Batch Processing**: Efficient handling of multiple addresses
        - **Progress Tracking**: Real-time status updates
        """)
    
    with impl_col2:
        st.markdown("""
        ### AI Integration:
        - **Prompt Engineering**: Optimised LLM prompts for accuracy
        - **Chain Processing**: Sequential AI analysis stages
        - **Error Handling**: Robust failure recovery mechanisms
        - **Result Validation**: Quality assurance checks
        """)
    
    # Footer
    st.markdown("---")
    st.markdown("""
    <div style='text-align: center; color: #666666; font-style: italic;'>
    Methodology Documentation - Technical Implementation Guide
    </div>
    """, unsafe_allow_html=True)


if __name__ == "__main__":
    display_methodology_page()
