requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Django==5.2.4
gitdb==4.0.12
GitPython==3.1.41
google_search_results==2.4.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
image==1.5.33
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.5
MarkupSafe==3.0.2
narwhals==1.46.0
numpy==2.3.1
openai==1.95.1
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
plotly==5.24.1
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
serpapi==0.1.5
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
streamlit==1.46.1
streamlit-chat==0.1.1
streamlit-timeline==0.0.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zstandard==0.23.0
